+++ bfabric_scripts/docs/changelog.md
@@ -14,6 +14,7 @@ Versioning currently follows `X.Y.Z` where
 
 - `bfabric-cli executable dump` command to export YAML and XML of executables
 - `bfabric-cli feeder create-importresource` command to register importresource objects in B-Fabric
+- `bfabric-cli executable upload` supports XML
 
 ### Changed